+++ extensions/reef/tsconfig.json
@@ -3,10 +3,14 @@
   "compilerOptions": {
     "rootDir": "."
   },
-  "include": ["./*.ts", "./src/**/*.ts", "./protocol/**/*.ts"],
+  "include": ["./*.ts", "./src/**/*.ts"],
   "exclude": [
     "./**/*.test.ts",
     "./dist/**",
-    "./node_modules/**"
+    "./node_modules/**",
+    "./src/test-support/**",
+    "./src/**/*test-helpers.ts",
+    "./src/**/*test-harness.ts",
+    "./src/**/*test-support.ts"
   ]
 }
